Move PBox into palloc::PBox
