+++ pgrx/src/memcx.rs
@@ -10,8 +10,7 @@ use pgrx_sql_entity_graph::metadata::{
 // - mctx
 // Search engines will see "memc[tx]{2}" and assume you mean memcpy!
 // And it's nice-ish to have shorter lifetime names and have 'mcx consistently mean the lifetime.
-use crate::callconv::{Arg, ArgAbi, BoxRet, FcInfo};
-use crate::datum::{BorrowDatum, Datum};
+use crate::callconv::{Arg, ArgAbi};
 use crate::pg_sys;
 use core::{marker::PhantomData, ptr::NonNull};
 
@@ -150,40 +149,5 @@ unsafe impl<'mcx> SqlTranslatable for &MemCx<'mcx> {
 
     fn return_sql() -> Result<Returns, ReturnsError> {
         Ok(Returns::One(SqlMapping::Skip))
-    }
-}
-
-/// A type pallocated in a context
-pub struct PBox<'mcx, T: ?Sized> {
-    ptr: NonNull<T>,
-    _cx: PhantomData<MemCx<'mcx>>,
-}
-
-impl<'mcx, T: ?Sized> PBox<'mcx, T> {
-    pub unsafe fn from_raw_in(ptr: NonNull<T>, _cx: &MemCx<'mcx>) -> PBox<'mcx, T> {
-        PBox { ptr, _cx: PhantomData }
-    }
-}
-
-unsafe impl<'mcx, T> BoxRet for PBox<'mcx, T>
-where
-    T: ?Sized + BorrowDatum,
-{
-    unsafe fn box_into<'fcx>(self, fcinfo: &mut FcInfo<'fcx>) -> Datum<'fcx> {
-        // SAFETY: by proxy
-        unsafe { fcinfo.return_raw_datum(pg_sys::Datum::from(self.ptr.cast::<u8>().as_ptr())) }
-    }
-}
-
-unsafe impl<'mcx, T> SqlTranslatable for PBox<'mcx, T>
-where
-    T: SqlTranslatable + ?Sized,
-{
-    fn argument_sql() -> Result<SqlMapping, ArgumentError> {
-        T::argument_sql()
-    }
-
-    fn return_sql() -> Result<Returns, ReturnsError> {
-        T::return_sql()
     }
 }
