+++ pgrx/src/palloc.rs
@@ -1,0 +1,3 @@
+mod pbox;
+
+pub use pbox::PBox;
